+++ flower.common/src/test/java/com/ly/train/flower/common/actor/ServiceFacadeTest.java
@@ -26,6 +26,7 @@ import com.ly.train.flower.base.service.ServiceA;
 import com.ly.train.flower.base.service.ServiceB;
 import com.ly.train.flower.base.service.ServiceC1;
 import com.ly.train.flower.base.service.ServiceC2;
+import com.ly.train.flower.common.akka.ServiceFacade;
 import com.ly.train.flower.common.service.container.ServiceFlow;
 
 /**
