+++ flower.sample/src/main/java/com/ly/train/flower/common/sample/textflow/Sample.java
@@ -16,7 +16,7 @@
 package com.ly.train.flower.common.sample.textflow;
 
 import org.junit.Test;
-import com.ly.train.flower.common.actor.ServiceFacade;
+import com.ly.train.flower.common.akka.ServiceFacade;
 import com.ly.train.flower.common.sample.textflow.model.Message1;
 import com.ly.train.flower.common.sample.textflow.model.Message2;
