+++ flower.sample/src/main/java/com/ly/train/flower/common/sample/textflow/ServiceRouterSample.java
@@ -16,8 +16,8 @@
 package com.ly.train.flower.common.sample.textflow;
 
 import org.junit.Test;
-import com.ly.train.flower.common.actor.ServiceFacade;
-import com.ly.train.flower.common.actor.ServiceRouter;
+import com.ly.train.flower.common.akka.ServiceFacade;
+import com.ly.train.flower.common.akka.ServiceRouter;
 import com.ly.train.flower.common.sample.textflow.model.Message1;
 import com.ly.train.flower.common.sample.textflow.model.Message2;
