+++ flower.sample/src/main/java/com/ly/train/flower/common/sample/web/forktest/BlockServlet.java
@@ -15,16 +15,16 @@
  */
 package com.ly.train.flower.common.sample.web.forktest;
 
-import com.alibaba.fastjson.JSONObject;
-import com.ly.train.flower.common.sample.web.forktest.service.BlockService;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.support.ClassPathXmlApplicationContext;
+import java.io.IOException;
+import java.io.PrintWriter;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.io.PrintWriter;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.support.ClassPathXmlApplicationContext;
+import com.alibaba.fastjson.JSONObject;
+import com.ly.train.flower.common.sample.web.forktest.service.BlockService;
 
 /**
  * @author fengyu.zhang
